DPC-837: Address code review comments
Rename factory classes to be more consistent and clear as to what each class actually does/provides.
