+++ dpc-common/src/main/java/gov/cms/dpc/fhir/parameters/ProvenanceResourceValueFactory.java
@@ -2,43 +2,51 @@ package gov.cms.dpc.fhir.parameters;
 
 import ca.uhn.fhir.context.FhirContext;
 import com.google.inject.Injector;
-import gov.cms.dpc.fhir.annotations.ProvenanceHeader;
 import org.glassfish.hk2.api.Factory;
-import org.glassfish.jersey.server.model.Parameter;
-import org.glassfish.jersey.server.spi.internal.ValueFactoryProvider;
-import org.hl7.fhir.instance.model.api.IBaseResource;
+import org.hl7.fhir.dstu3.model.Provenance;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
-import javax.ws.rs.ext.Provider;
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 
 /**
- * Custom {@link ValueFactoryProvider} that lets us cleanly extract {@link org.hl7.fhir.dstu3.model.Provenance} resources from the {@link ProvenanceResourceFactory#PROVENANCE_HEADER}.
+ * {@link Factory} for extracting a {@link org.hl7.fhir.dstu3.model.Provenance} resource from a request header.
  */
-@Provider
-public class ProvenanceResourceValueFactory implements ValueFactoryProvider {
+public class ProvenanceResourceValueFactory implements Factory<Provenance> {
+
+    static final String PROVENANCE_HEADER = "X-Provenance";
+    private static final Logger logger = LoggerFactory.getLogger(ProvenanceResourceValueFactory.class);
 
     private final Injector injector;
     private final FhirContext ctx;
 
-    @Inject
-    public ProvenanceResourceValueFactory(Injector injector, FhirContext ctx) {
+    ProvenanceResourceValueFactory(Injector injector, FhirContext ctx) {
         this.injector = injector;
         this.ctx = ctx;
     }
 
     @Override
-    public Factory<?> getValueFactory(Parameter parameter) {
-        if (parameter.getDeclaredAnnotation(ProvenanceHeader.class) != null) {
-            // If the parameter is a resource, pass it off to the resource factory
-            if (IBaseResource.class.isAssignableFrom(parameter.getRawType()))
-                return new ProvenanceResourceFactory(injector, ctx);
+    public Provenance provide() {
+        final HttpServletRequest request = injector.getInstance(HttpServletRequest.class);
+        final String headerValue = request.getHeader(PROVENANCE_HEADER);
+        if (headerValue == null) {
+            throw new WebApplicationException(String.format("Must have %s header", PROVENANCE_HEADER), Response.Status.BAD_REQUEST);
+        }
+        final Provenance provenance;
+        try {
+            provenance = ctx.newJsonParser().parseResource(Provenance.class, headerValue);
+        } catch (Exception e) {
+            logger.error("Cannot parse Provenance", e);
+            throw new WebApplicationException("Cannot parse FHIR `Provenance` resource", Response.Status.BAD_REQUEST);
         }
 
-        return null;
+        return provenance;
     }
 
     @Override
-    public PriorityType getPriority() {
-        return Priority.NORMAL;
+    public void dispose(Provenance instance) {
+        // Not used
     }
 }
